+++ extensions/healthie/actions/createMetricEntry/createMetricEntry.ts
@@ -1,0 +1,79 @@
+import { validate, type Action } from '@awell-health/extensions-core'
+import { Category } from '@awell-health/extensions-core'
+import { z, ZodError } from 'zod'
+import { fromZodError } from 'zod-validation-error'
+import { HealthieError, mapHealthieToActivityError } from '../../errors'
+import { getSdk } from '../../gql/sdk'
+import { initialiseClient } from '../../graphqlClient'
+import { settingsValidationSchema, type settings } from '../../settings'
+import { fields, FieldsValidationSchema } from './config'
+
+export const createMetricEntry: Action<typeof fields, typeof settings> = {
+  key: 'createMetricEntry',
+  category: Category.EHR_INTEGRATIONS,
+  title: 'Create metric entry',
+  description: 'Create a metric entry for a patient in Healthie.',
+  fields,
+  previewable: true,
+  onActivityCreated: async (payload, onComplete, onError): Promise<void> => {
+    try {
+      const {
+        settings,
+        fields: { userId, category, metricStat },
+      } = validate({
+        schema: z.object({
+          settings: settingsValidationSchema,
+          fields: FieldsValidationSchema,
+        }),
+        payload,
+      })
+
+      const client = initialiseClient(settings)
+      if (client != null) {
+        const sdk = getSdk(client)
+        await sdk.createEntry({
+          metric_stat: String(metricStat),
+          user_id: userId,
+          type: 'MetricEntry',
+          category,
+        })
+        await onComplete({})
+      }
+    } catch (err) {
+      if (err instanceof ZodError) {
+        const error = fromZodError(err)
+        await onError({
+          events: [
+            {
+              date: new Date().toISOString(),
+              text: { en: error.message },
+              error: {
+                category: 'WRONG_INPUT',
+                message: error.message,
+              },
+            },
+          ],
+        })
+      } else if (err instanceof HealthieError) {
+        const errors = mapHealthieToActivityError(err.errors)
+        await onError({
+          events: errors,
+        })
+      } else {
+        const error = err as Error
+        await onError({
+          events: [
+            {
+              date: new Date().toISOString(),
+              text: { en: 'Healthie API reported an error' },
+              error: {
+                category: 'SERVER_ERROR',
+                message: error.message,
+              },
+            },
+          ],
+        })
+      }
+    }
+  },
+}